Tons of refactoring and reformulating. Lots of work on I/O capabilities.
