+++ sandbox/example/.gitignore
@@ -1,4 +1,5 @@
-gnds
+01
+02
+03
 pugixml.o
-t
 collate
