+++ sandbox/example/compile-clang++
@@ -1,0 +1,30 @@
+#!/bin/bash
+
+# ------------------------
+# Setup
+# ------------------------
+
+COMPILE='
+   clang++
+  -std=c++17
+  -I../../subprojects/knoop/src
+  -I../../subprojects/range-v3-adapter/src/include
+  -I../../subprojects/valuable-adapter/src/valuable/include
+  -I../../subprojects/pugixml-adapter/src/src
+  -I../../subprojects/json-adapter/src/src
+  -I../src
+  -O3 -Wall -Wextra -Wpedantic '
+
+LIBRARIES=''
+
+# ------------------------
+# Compile
+# ------------------------
+
+if [ ! -f "pugixml.o" ]; then
+$COMPILE ../../subprojects/pugixml-adapter/src/src/pugixml.cpp -c -o pugixml.o
+fi
+
+$COMPILE example-01.cpp pugixml.o -o 01 $LIBRARIES
+$COMPILE example-02.cpp pugixml.o -o 02 $LIBRARIES
+$COMPILE example-03.cpp pugixml.o -o 03 $LIBRARIES
