+++ sandbox/misc/collate.cpp
