+++ sandbox/misc/collate.sh
